refactor: consolidate utility imports for competition and course data across components -getCourseListをAPIからではなく直でDBから取得
